feat: Move EventManager to the new service
