+++ common/python/powerpi_common/container.py
@@ -5,7 +5,6 @@ from powerpi_common.condition import ConditionContainer
 from powerpi_common.config.config_retriever import ConfigRetriever
 from powerpi_common.controller import Controller
 from powerpi_common.device import DeviceContainer
-from powerpi_common.event import EventManager
 from powerpi_common.health import HealthService
 from powerpi_common.logger import Logger
 from powerpi_common.mqtt import MQTTClient
@@ -71,15 +70,6 @@ class Container(containers.DeclarativeContainer):
         variable_manager=variable.variable_manager
     )
 
-    event_manager = providers.Singleton(
-        EventManager,
-        config=config,
-        logger=logger,
-        mqtt_client=mqtt_client,
-        device_manager=device.device_manager,
-        variable_manager=variable.variable_manager
-    )
-
     health = providers.Singleton(
         HealthService,
         config=config,
@@ -93,7 +83,6 @@ class Container(containers.DeclarativeContainer):
         logger=logger,
         config_retriever=config_retriever,
         device_manager=device.device_manager,
-        event_manager=event_manager,
         mqtt_client=mqtt_client,
         device_status_checker=device.device_status_checker,
         scheduler=scheduler,
