+++ services/event/event/config.py
@@ -5,4 +5,4 @@ from powerpi_common.config import ConfigFileType
 class EventConfig(CommonConfig):
     @property
     def used_config(self):
-        return [ConfigFileType.EVENTS]
+        return [ConfigFileType.DEVICES, ConfigFileType.EVENTS]
